requirements.txt
pytest~=7.2.0
pytest-html
pytest-xdist
pytest-ordering
pytest-rerunfailures
allure-pytest
pytest-base-url
PyYAML~=6.0
requests~=2.28.1